+++ aws_teams_logger/loggers/__init__.py
@@ -1,10 +1,12 @@
 __all__ = [
     'LambdaLogger',
     'TaskLogger',
+    'BulkLambdaLogger',
+    'BulkTaskLogger',
     'set_account_name'
 ]
 
 
-from .lambda_logger import LambdaLogger
-from .task_logger import TaskLogger
+from .lambda_logger import LambdaLogger, BulkLambdaLogger
+from .task_logger import TaskLogger, BulkTaskLogger
 from .globals import set_account_name