+++ aws_teams_logger/utils/aws/ses/ses.py
@@ -1,10 +1,13 @@
 import os
+from concurrent.futures import ThreadPoolExecutor
 from json import dumps
 from typing import Union, List, Dict, Any
 
 from aws_teams_logger.log import LOG
+from .models import BulkDestination
 from ..client_cache import ClientCache
 from ...decorators import log_time
+from ...split import divide_chunks
 from ...types import as_list
 
 
@@ -12,14 +15,18 @@ class SESHelper(ClientCache):
 
     SERVICE_NAME = 'ses'
 
+    # There is a hard limit of 50 Destinations when making a call to the
+    # `ses:SendBulkTemplatedEmail` API
+    MAX_BULK_DESTINATIONS = 50
+
     @staticmethod
     def _op_is_success(response: Dict[str, Any]) -> bool:
         """Check if response indicates an operation was a success"""
         return response.get('ResponseMetadata', {}).get('HTTPStatusCode') == 200
 
     @log_time
-    def send_templated_email(self,
-                             name: str, data: Dict[str, Any],
+    def send_templated_email(self, name: str,
+                             data: Dict[str, Any],
                              to_addresses: Union[List[str], str],
                              from_address: str,
                              reply_to_addresses: Union[List[str], str]):
@@ -35,6 +42,55 @@ class SESHelper(ClientCache):
         )
 
         return response
+
+    @log_time
+    def send_bulk_templated_email(self, name: str,
+                                  destinations: List[BulkDestination],
+                                  from_address: str,
+                                  reply_to_addresses: Union[List[str], str]):
+        """
+        Send a bulk templated email via SES client
+
+        Ref: https://docs.aws.amazon.com/ses/latest/APIReference/API_SendBulkTemplatedEmail.html
+        """
+        if not destinations:
+            return
+
+        destination_list = [dest.dict for dest in destinations]
+
+        if len(destination_list) > self.MAX_BULK_DESTINATIONS:
+            # We have more than :attr:`MAX_BULK_DESTINATIONS` emails to send.
+            #
+            # AWS recommends dividing the recipient list into groups of 50 or
+            # fewer, and calling the SendBulkTemplatedEmail operation several
+            # times on each group.
+            dest_chunks = list(divide_chunks(
+                destination_list, self.MAX_BULK_DESTINATIONS))
+            num_workers = min(10, len(dest_chunks))
+
+            with ThreadPoolExecutor(max_workers=num_workers) as pool:
+                for dest_chunk in dest_chunks:
+                    pool.submit(self._call_bulk_templated_email,
+                                name, dest_chunk,
+                                from_address, reply_to_addresses)
+
+        else:
+            # We have fewer than :attr:`MAX_BULK_DESTINATIONS` emails - it can
+            # be sent in a single API call.
+            self._call_bulk_templated_email(
+                name, destination_list, from_address, reply_to_addresses)
+
+    def _call_bulk_templated_email(
+            self, name: str, destination_list: List[Dict],
+            from_address: str, reply_to_addresses: Union[List[str], str]):
+        """Calls the underlying `ses:SendBulkTemplatedEmail` operation"""
+        return self.client.send_bulk_templated_email(
+            Source=from_address,
+            Destinations=destination_list,
+            ReplyToAddresses=as_list(reply_to_addresses),
+            Template=name,
+            DefaultTemplateData=dumps(destination_list[0])
+        )
 
     def replace_template(self, template_data: Dict[str, Any]) -> bool:
         """